devops: auto deploy lambda
