+++ .github/workflows/lambda-deploy.yml
@@ -1,0 +1,36 @@
+name: Deploy Lambda Function
+
+on:
+  push:
+    branches:
+      - develop
+
+jobs:
+  deploy:
+    runs-on: ubuntu-latest
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v2
+
+    - name: Set up Node.js
+      uses: actions/setup-node@v2
+      with:
+        node-version: '20'
+
+    - name: Install dependencies
+      run: |
+        cd lambda
+        npm install
+
+    - name: Archive project
+      run: |
+        cd lambda
+        zip -r ../lambda.zip blackheathWeatherLambda.js node_modules package.json package-lock.json
+
+    - name: Deploy to AWS Lambda
+      env:
+        AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
+        AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
+        AWS_REGION: ap-southeast-2
+      run: aws lambda update-function-code --function-name weatherStation --zip-file fileb://lambda.zip
